+++ ErsatzTV.Application/MediaCards/ImageCardResultsViewModel.cs
@@ -1,0 +1,5 @@
+﻿using ErsatzTV.Core.Search;
+
+namespace ErsatzTV.Application.MediaCards;
+
+public record ImageCardResultsViewModel(int Count, List<ImageCardViewModel> Cards, SearchPageMap PageMap);