+++ ErsatzTV.Application/MediaCollections/Commands/AddItemsToCollection.cs
@@ -12,4 +12,5 @@ public record AddItemsToCollection
     List<int> ArtistIds,
     List<int> MusicVideoIds,
     List<int> OtherVideoIds,
-    List<int> SongIds) : IRequest<Either<BaseError, Unit>>;
+    List<int> SongIds,
+    List<int> ImageIds) : IRequest<Either<BaseError, Unit>>;