+++ ErsatzTV.Core/Domain/Metadata/ImageMetadata.cs
@@ -1,0 +1,8 @@
+﻿namespace ErsatzTV.Core.Domain;
+
+public class ImageMetadata : Metadata
+{
+    public int? DurationSeconds { get; set; }
+    public int ImageId { get; set; }
+    public Image Image { get; set; }
+}