+++ ErsatzTV.Infrastructure/Data/TvContext.cs
@@ -53,6 +53,8 @@ public class TvContext : DbContext
     public DbSet<OtherVideoMetadata> OtherVideoMetadata { get; set; }
     public DbSet<Song> Songs { get; set; }
     public DbSet<SongMetadata> SongMetadata { get; set; }
+    public DbSet<Image> Images { get; set; }
+    public DbSet<ImageMetadata> ImageMetadata { get; set; }
     public DbSet<Show> Shows { get; set; }
     public DbSet<ShowMetadata> ShowMetadata { get; set; }
     public DbSet<Season> Seasons { get; set; }