+++ ErsatzTV.Infrastructure/Images/ImageCache.cs
@@ -12,6 +12,7 @@ using ErsatzTV.Core.Interfaces.Metadata;
 using SixLabors.ImageSharp;
 using SixLabors.ImageSharp.PixelFormats;
 using SixLabors.ImageSharp.Processing;
+using Image=SixLabors.ImageSharp.Image;
 
 namespace ErsatzTV.Infrastructure.Images;