+++ ErsatzTV/Pages/MultiSelectBase.cs
@@ -18,7 +18,7 @@ public class MultiSelectBase<T> : FragmentNavigationBase
     public MultiSelectBase()
     {
         _recentlySelected = None;
-        SelectedItems = new System.Collections.Generic.HashSet<MediaCardViewModel>();
+        SelectedItems = [];
     }
 
     [Inject]
@@ -89,7 +89,8 @@ public class MultiSelectBase<T> : FragmentNavigationBase
         SelectedItems.OfType<ArtistCardViewModel>().Map(a => a.ArtistId).ToList(),
         SelectedItems.OfType<MusicVideoCardViewModel>().Map(mv => mv.MusicVideoId).ToList(),
         SelectedItems.OfType<OtherVideoCardViewModel>().Map(ov => ov.OtherVideoId).ToList(),
-        SelectedItems.OfType<SongCardViewModel>().Map(s => s.SongId).ToList());
+        SelectedItems.OfType<SongCardViewModel>().Map(s => s.SongId).ToList(),
+        SelectedItems.OfType<ImageCardViewModel>().Map(i => i.ImageId).ToList());
 
     protected async Task AddItemsToCollection(
         List<int> movieIds,
@@ -100,6 +101,7 @@ public class MultiSelectBase<T> : FragmentNavigationBase
         List<int> musicVideoIds,
         List<int> otherVideoIds,
         List<int> songIds,
+        List<int> imageIds,
         string entityName = "selected items")
     {
         int count = movieIds.Count + showIds.Count + seasonIds.Count + episodeIds.Count + artistIds.Count +
@@ -123,7 +125,8 @@ public class MultiSelectBase<T> : FragmentNavigationBase
                 artistIds,
                 musicVideoIds,
                 otherVideoIds,
-                songIds);
+                songIds,
+                imageIds);
 
             Either<BaseError, Unit> addResult = await Mediator.Send(request, CancellationToken);
             addResult.Match(